% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{amed}
\alias{amed}
\title{Average Marginal Effects}
\usage{
amed(model)
}
\arguments{
\item{model}{the model object}
}
\description{
Internal function for \code{mma()}. Based on the same strategy as 
\code{margins} by T. Leeper.
}
\author{
Tyson S. Barrett
}
