% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Piping operator re-exported from \code{magrittr}}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{left-hand side}

\item{rhs}{right-hand side}
}
\description{
Piping operator re-exported from \code{magrittr}
}
