% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marginal_Mediation.R
\name{mma}
\alias{mma}
\title{Marginal Mediation}
\usage{
mma(data, ..., family, ind_effects, boot = 100, ci = 0.95)
}
\arguments{
\item{data}{the data frame with the data for the models}

\item{...}{formulas for the models; the first is the model with the outcome while the others are the mediated effects ("a" paths)}

\item{family}{the vector of the families of the model. Either \code{binomial} for binary outcomes or \code{gaussian} for continuous. Needs to have a length equal to the number of models.}

\item{ind_effects}{a vector of the desired indirect effects. Has the form \code{"var1-var2"}.}

\item{boot}{the number of bootstrapped samples; default is 100}

\item{ci}{the confidence interval; the default is .975 which is the 95\% confidence interval.}
}
\description{
Provides the ability to perform marginal mediation. Marginal mediation
is particularly useful for situations where the mediator or outcome is dichotomous.
The results provide the average marginal effects of the models, providing simple
interpretation of the indirect effects.
}
\details{
Using the average marginal effects as discussed by Tamas Bartus (2005), 
the coefficients are transformed into probabilities (for binary outcomes) or remain
in their original units (continuous outcomes).
}
\references{
Bartus, T. (2005). Estimation of marginal effects using margeff. 
The Stata Journal, 5(3), 309–329.
}
\author{
Tyson S. Barrett
}
