% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checks.R
\name{mma_check}
\alias{mma_check}
\title{Uncorrelated Residual Assumption Check}
\usage{
mma_check(model)
}
\arguments{
\item{model}{The mma model object}
}
\description{
Provides the correlations of the residual terms of the model
}
