% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comps.R
\name{mma_dir_effects}
\alias{mma_dir_effects}
\title{Direct Effects Extraction for MMA}
\usage{
mma_dir_effects(model)
}
\arguments{
\item{model}{mma fit object}
}
\description{
Extracts the formulas from a mma object
}
