% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marginal_Mediation.R
\name{pdfed}
\alias{pdfed}
\title{PDF of Logistic and Gaussian Models}
\usage{
pdfed(x, family)
}
\arguments{
\item{x}{the model}

\item{family}{family of the model}
}
\description{
Internal function for \code{margins_mediation()}.
}
\author{
Tyson S. Barrett
}
