% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comps.R
\name{perc_med}
\alias{perc_med}
\title{Percent Mediation}
\usage{
perc_med(model, effect)
}
\arguments{
\item{model}{mma fit object}

\item{effect}{the indirect effect to be compared to its direct path}
}
\description{
To obtain the percent of the total effect that is mediated through the specified
indirect path: indirect / (total) * 100.
}
