% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comps.R
\name{rel_effects}
\alias{rel_effects}
\title{Indirect vs Direct Comparison}
\usage{
rel_effects(model, variable)
}
\arguments{
\item{model}{mma fit object}

\item{variable}{the variable in the model that is to be compared}
}
\description{
To make comparisons between indirect and direct effects of a variable.
The default comparison is indirect / direct.
}
