% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_mod}
\alias{run_mod}
\title{Bootstrapped Models}
\usage{
run_mod(data, indices, formula, family)
}
\arguments{
\item{data}{the data.frame with the data for the models.}

\item{indices}{the indices for each bootstrapped model (internal in \code{boot()})}

\item{formula}{the formula of the model}

\item{family}{the family of the model}
}
\description{
Internal function for \code{margins_mediation()}.
}
\author{
Tyson S. Barrett
}
