% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marginal_Mediation.R
\name{run_mods}
\alias{run_mods}
\title{Bootstrapped Models}
\usage{
run_mods(data, mods, family, boot = 100)
}
\arguments{
\item{data}{the data.frame with the data for the models.}

\item{mods}{the formulas for the models.}

\item{family}{the family of the GLM model.}

\item{boot}{the number of bootstrapped samples.}
}
\description{
Internal function for \code{margins_mediation()}.
}
\author{
Tyson S. Barrett
}
