% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marginal_Mediation.R
\name{mma}
\alias{mma}
\title{Marginal Mediation}
\usage{
mma(data, ..., family, ind_effects, ci_type = "perc", boot = 100,
  ci = 0.95)
}
\arguments{
\item{data}{the data frame with the data for the models}

\item{...}{formulas for the models; the first is the model with the outcome while the others are the mediated effects ("a" paths)}

\item{family}{the vector of the families of the model. Either \code{binomial} for binary outcomes or \code{gaussian} for continuous. Needs to have a length equal to the number of models.}

\item{ind_effects}{a vector of the desired indirect effects. Has the form \code{"var1-var2"}.}

\item{ci_type}{a string indicating the type of bootstrap method to use (currently "perc" and "basic" are available; "perc" is recommended). Further development will allow the Bias-Corrected bootstrap soon.}

\item{boot}{the number of bootstrapped samples; default is 100}

\item{ci}{the confidence interval; the default is .975 which is the 95\% confidence interval.}
}
\value{
A list of class \code{mma} containing:
\item{ind_effects}{the indirect effects reported in the average marginal effect}
\item{dir_effects}{the direct effects reported in the average marginal effect}
\item{ci_level}{the confidence level}
\item{data}{the original data frame}
\item{reported_ind}{the indirect effects the user requested (in the \code{...})}
\item{boot}{the number of bootstrap samples}
\item{model}{the formulas of the individual sub-models}
\item{call}{the original function call}
}
\description{
Provides the ability to perform marginal mediation. Marginal mediation
is particularly useful for situations where the mediator or outcome is categorical,
a count, or some other non-normally distributed variable.
The results provide the average marginal effects of the models, providing simple
interpretation of the indirect effects.
}
\details{
Using the average marginal effects as discussed by Tamas Bartus (2005), 
the coefficients are transformed into probabilities (for binary outcomes) or remain
in their original units (continuous outcomes).
}
\examples{
\dontrun{
library(furniture)
data(nhanes_2010)
(fit = mma(nhanes_2010,
           marijuana ~ home_meals + gender + age + asthma,
           home_meals ~ gender + age + asthma,
           age ~ gender + asthma,
           family = c(binomial, gaussian, gaussian),
           ind_effects = c("genderFemale-home_meals",
                           "age-home_meals",
                           "asthmaNo-home_meals",
                           "genderFemale-age",
                           "asthmaNo-age"),
           boot = 500))
}

}
\references{
Bartus, T. (2005). Estimation of marginal effects using margeff. 
The Stata Journal, 5(3), 309–329.
}
\author{
Tyson S. Barrett
}
