% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdfed}
\alias{pdfed}
\title{PDF of logistic, probit, and poisson models}
\usage{
pdfed(model)
}
\arguments{
\item{model}{the model}
}
\description{
Internal function for \code{margins_mediation()}. More functionality to be added later.
}
\author{
Tyson S. Barrett
}
