% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comps.R
\name{rel_effects}
\alias{rel_effects}
\title{Indirect Effect vs Total Effect}
\usage{
rel_effects(model, effect)
}
\arguments{
\item{model}{mma fit object}

\item{effect}{the indirect effect to be compared to its direct path}
}
\description{
To make comparisons between indirect and direct effects of a variable.
The default comparison is indirect / (direct + indirect)
}
